+++ src/components/spartan/SRadio/index.ts
@@ -1,1 +1,3 @@
 export { default as SRadio } from './SRadio.vue';
+
+export type { TRadioEmits, TRadioProps } from './types';